+++ Sources/ProjectSpec/Yaml.swift
@@ -3,11 +3,17 @@ import PathKit
 import Yams
 
 public func loadYamlDictionary(path: Path) throws -> [String: Any] {
-    let string: String = try path.read()
-    if string == "" {
+    let data = try path.read()
+    if data.isEmpty {
         return [:]
     }
+    return try loadYamlDictionary(from: data)
+}
 
+func loadYamlDictionary(from data: Data) throws -> [String: Any] {
+    guard let string = String(data: data, encoding: .utf8) else {
+        return [:]
+    }
     let resolver = Resolver.default
         .removing(.null) // remove rule so that empty quotes are treated as empty strings