+++ CHANGELOG.md
@@ -105,6 +105,7 @@
 ## 2.11.0
 
 #### Added
+- Support optional include. [#712](https://github.com/yonaskolb/XcodeGen/pull/712) @giginet
 - Add Carthage static framework dependencies support. [#688](https://github.com/yonaskolb/XcodeGen/pull/688) @giginet
 - Added `xcodegen dump` command [#710](https://github.com/yonaskolb/XcodeGen/pull/710) @yonaskolb
 - Added `--no-env` option to disable environment variables expansion [#704](https://github.com/yonaskolb/XcodeGen/pull/704) @rcari